refactor(web): Phase 4 — lift cronjobs into 3-tier (Infra → Service → API)
Migrates cronjobs to the API → Services → Infra layout per
`docs/web-architecture.md` (issue #315):

- `server/infra/db/queries/cronjobs.ts` — `CronjobQueries` (extracted from
  `lib/cronjob-store.ts`).
- `server/services/cronjob-service.ts` — `CronjobService`, absorbing
  `lib/cronjob-scheduler.ts` (timer lifecycle, chat routing, fire path)
  + `lib/cronjob-store.ts` (persistence). Scheduler state held on a
  `globalThis` symbol to survive HMR. Exposes domain errors
  (`CronjobNotFoundError`, `InvalidCronExpressionError`, …) so the
  service tier stays framework-agnostic.
- `server/api/cronjobs/router.ts` — thin tRPC sub-router that delegates
  to `cronjobService` and translates domain errors into tRPC codes.

Removes the `cronjobs` sub-router from the legacy `trpc/router.ts` and
merges the new one in `server/api/router.ts`. Project/workspace delete
paths switch from the two-step `stopJobsForKey + deleteCronjobFile`
pair to the single `cronjobService.removeForKey(key)`. `start-server.ts`
swaps `startCronjobScheduler` / `stopCronjobScheduler` for the service's
`start()` / `stop()`.

The legacy `lib/cronjob-{scheduler,store,types}.ts` files are deleted.
Behaviour, error codes, and the `band:cronId`-labeled chat reuse from
`apps/web/tests/cronjobs.test.ts` suite (19 tests, all green).

Closes #315
