+++ apps/web/src/server/api/router.ts
@@ -9,6 +9,7 @@
  * Migration order:
  *   - Phase 1 (issue #312): `settings/`.
  *   - Phase 2 (issue #313): `projects/`.
+ *   - Phase 4 (issue #315): `cronjobs/`.
  *
  * The rest of the procedures still live in the legacy
  * `apps/web/src/trpc/router.ts`; we merge that legacy router with every
@@ -21,17 +22,18 @@
  */
 
 import { appRouter as legacyAppRouter } from "../../trpc/router";
+import { cronjobsRouter } from "./cronjobs/router";
 import { projectsRouter } from "./projects/router";
 import { settingsRouter } from "./settings/router";
 import { t } from "./trpc";
 
 // INVARIANT: the legacy router (`apps/web/src/trpc/router.ts`) must not
-// contain any key that this file also defines (`settings`, `projects`, …).
-// `t.mergeRouters` accepts two routers and silently picks last-write-wins
-// for duplicate keys, so a stray legacy entry would mask the migrated
-// router without a build error. Each phase of the 3-tier migration adds
-// a key here and removes it from the legacy router in the same diff;
-// the invariant must hold for every key composed below.
+// contain any key that this file also defines (`settings`, `projects`,
+// `cronjobs`, …). `t.mergeRouters` accepts two routers and silently picks
+// last-write-wins for duplicate keys, so a stray legacy entry would mask
+// the migrated router without a build error. Each phase of the 3-tier
+// migration adds a key here and removes it from the legacy router in the
+// same diff; the invariant must hold for every key composed below.
 //
 // Live guards:
 //   - `tRPC — settings CRUD` in `apps/web/tests/trpc.test.ts` exercises
@@ -42,11 +44,15 @@ import { t } from "./trpc";
 //     `apps/web/tests/plain-projects.test.ts` exercise the projects
 //     sub-router, so a duplicate-key regression on `projects.*` trips
 //     them via the same path.
+//   - `apps/web/tests/cronjobs.test.ts` exercises `cronjobs.*` end-to-end,
+//     so a regression that masks the migrated `cronjobsRouter` with a
+//     stale legacy entry trips one of those assertions.
 export const appRouter = t.mergeRouters(
   legacyAppRouter,
   t.router({
     settings: settingsRouter,
     projects: projectsRouter,
+    cronjobs: cronjobsRouter,
   }),
 );
