+++ docs/cli.md
@@ -18,7 +18,7 @@ There are two reasons to use them:
 1. **Organize your own chats.** Tag a chat with `phase=plan` / `phase=implement` / `phase=review`, or by feature area, owner, or anything else. Filter the list client-side with `band chats list --output json | jq '.chats[] | select(.labels.phase == "plan")'`.
 2. **Let the cronjob scheduler claim its own chat.** Each cronjob owns a dedicated chat in the target workspace, identified by the reserved `band:cronId` label. On the first fire the scheduler creates a chat tagged with `band:cronId=<jobId>`; subsequent fires reuse the same chat by looking it up via `findChatByLabels`. The user can delete the chat at any time — the next fire will recreate it (intentional soft reset). This replaces the older "dispatch to whichever chat happens to be active" behaviour, so cron output no longer interleaves with the user's interactive conversation.
 
-The `band:` key prefix is reserved for server-internal labels. Writes through the CLI or any user-facing tRPC route are rejected; only server code (e.g. `cronjob-scheduler.ts`) can set them, and it does so by passing `allowReservedLabels: true` to `createChat`. Other validation rules at the write boundary: at most 20 keys per chat, keys match `^[a-zA-Z0-9_:-]{1,64}$` (colons allowed for namespacing), values are non-empty printable ASCII up to 256 chars.
+The `band:` key prefix is reserved for server-internal labels. Writes through the CLI or any user-facing tRPC route are rejected; only server code (e.g. `server/services/cronjob-service.ts`) can set them, and it does so by passing `allowReservedLabels: true` to `createChat`. Other validation rules at the write boundary: at most 20 keys per chat, keys match `^[a-zA-Z0-9_:-]{1,64}$` (colons allowed for namespacing), values are non-empty printable ASCII up to 256 chars.
 
 ### CLI commands
